Move aws_kms AssumeRole and WebIdentity into it's own file
